+++ tests/test_core.py
@@ -1,34 +1,108 @@
+# coding: utf-8
+
+import itertools
+import sys
 import time
 
-import stbt
+import mock
+import pytest
+
+from stbt import wait_until
 
 
-def test_that_wait_until_returns_on_success():
-    count = [0]
-
-    def t():
-        count[0] += 1
-        return True
-
-    assert stbt.wait_until(t)
-    assert count == [1]
+# pylint:disable=redefined-outer-name,unused-argument
 
 
-def test_that_wait_until_times_out():
-    start = time.time()
-    assert not stbt.wait_until(lambda: False, timeout_secs=0.1)
-    end = time.time()
-    assert 0.1 < end - start < 0.2
+class f(object):
+    """Helper factory for wait_until selftests. Creates a callable object that
+    returns the specified values one by one each time it is called.
+
+    Values are specified as space-separated characters. `.` means `None` and
+    `F` means `False`.
+    """
+
+    mapping = {".": None, "F": False}
+
+    def __init__(self, spec):
+        self.spec = spec
+        self.iterator = itertools.cycle(f.mapping.get(x, x)
+                                        for x in self.spec.split())
+
+    def __repr__(self):
+        return "f(%r)" % self.spec
+
+    def __call__(self):
+        time.sleep(1)
+        v = next(self.iterator)
+        sys.stderr.write("f() -> %s\n" % v)
+        return v
 
 
-def test_that_wait_until_tries_one_last_time():
-    def t():
-        time.sleep(0.2)
-        yield False
-        yield True
+@pytest.fixture(scope="function")
+def mock_time():
+    """Mocks out `time.time()` and `time.sleep()` so that time only advances
+    when you call `time.sleep()`.
+    """
 
-    assert stbt.wait_until(t().next, timeout_secs=0.1)
+    t = [1497000000]
+
+    def _time():
+        sys.stderr.write("time.time() -> %s\n" % t[0])
+        return t[0]
+
+    def _sleep(n):
+        sys.stderr.write("time.sleep(%s)\n" % n)
+        t[0] += n
+
+    with mock.patch("time.time", _time), mock.patch("time.sleep", _sleep):
+        yield
 
 
-def test_that_wait_until_with_zero_timeout_tries_once():
-    assert stbt.wait_until(lambda: True, timeout_secs=0)
+class Zero(object):
+    def __nonzero__(self):
+        return False
+
+    def __eq__(self, other):
+        return isinstance(other, Zero)
+
+
+@pytest.mark.parametrize("f,kwargs,expected", [
+    # wait_until returns on success
+    (f(". a b"), {}, "a"),
+    (f("F a b"), {}, "a"),
+
+    # wait_until tries one last time after reaching timeout_secs
+    (f("F T"), {"timeout_secs": 1}, "T"),
+    (f("F T"), {"timeout_secs": 0.1}, "T"),
+    (f("F F T"), {"timeout_secs": 1}, False),
+
+    # wait_until with zero timeout tries once
+    (lambda: True, {"timeout_secs": 0}, True),
+
+    # stable_secs behaviour
+    (f("a b b"), {}, "a"),
+    (f("a b b"), {"stable_secs": 1}, "b"),
+    (f("a b c"), {"stable_secs": 1}, None),
+
+    # timeout_secs elapsed
+    #        |   ┌ stable_secs needed
+    #        v   v
+    (f("a b b b b b b b b b b"), {"timeout_secs": 3, "stable_secs": 3}, None),
+
+    # Timeout reached
+    #        | ┌ stable_secs reached
+    #        v v
+    (f("a b b b a a a a a a a"), {"timeout_secs": 3, "stable_secs": 2}, "b"),
+
+    # Falsey values
+    (f("F F F"), {}, False),
+    (f("F F F"), {"stable_secs": 1}, None),
+    (Zero, {"interval_secs": 1}, Zero())
+])
+def test_wait_until(mock_time, f, kwargs, expected):
+    assert wait_until(f, **kwargs) == expected
+
+
+def test_that_wait_until_times_out(mock_time):
+    assert not wait_until(Zero, interval_secs=1)
+    assert time.time() == 1497000010